+++ cert_chain_resolver/resolver.py
@@ -1,6 +1,7 @@
 from contextlib import closing
 
 from cert_chain_resolver.models import CertificateChain, Cert
+from cert_chain_resolver.utils import load_bytes_to_x509_all
 
 try:
     from urllib.request import urlopen, Request
@@ -51,12 +52,19 @@ def resolve(bytes_cert, _chain=None, root_ca_store=None):
 
     _chain += cert
 
-    parent_cert = None
+    parent_bytes = None
     if cert.ca_issuer_access_location:
-        parent_cert = _download(cert.ca_issuer_access_location)
+        bundle_bytes = _download(cert.ca_issuer_access_location)
+        parents = load_bytes_to_x509_all(bundle_bytes)
+        primary = Cert(parents[0])
+        for extra_x509 in parents[1:]:
+            extra = Cert(extra_x509)
+            if extra.is_cross_sign_of(primary):
+                _chain.add_cross_sign(extra)
+        parent_bytes = bundle_bytes
 
-    if parent_cert:
-        return resolve(parent_cert, _chain=_chain, root_ca_store=root_ca_store)
+    if parent_bytes:
+        return resolve(parent_bytes, _chain=_chain, root_ca_store=root_ca_store)
     elif not _chain.root and root_ca_store:
         _chain += root_ca_store.find_issuer(cert)